+++ src/lib/validation/utils/assignHandlersForInputs.js
@@ -22,6 +22,11 @@ function assignHandlersForSingleInput (input) {
         const comparisonType = condition.indexOf('!==') > -1 ? 'isNotEqual' : 'isEqual'
         const matchers = condition.split(comparisonType === 'isNotEqual' ? '!==' : '===')
         const validationContainer = input.closest('[data-dp-validate]')
+        /**
+           Use the form as the validation container. This is necessary because in some cases, such as wizard,
+           the value of the 'data-dp-validate-if' attribute may refer to a different fieldset.
+           By using the form, this condition can be found throughout the entire form.
+        */
         const form = validationContainer.tagName === 'FIELDSET' && validationContainer.form ? validationContainer.form : validationContainer
 
         try {