+++ src/lib/validation/utils/helpers.js
@@ -31,6 +31,11 @@ function validateConditionsFulfilled (input) {
     const comparisonType = condition.indexOf('!==') > -1 ? 'isNotEqual' : 'isEqual'
     const matchers = condition.split(comparisonType === 'isNotEqual' ? '!==' : '===')
     const validationContainer = input.closest('[data-dp-validate]')
+    /**
+     Use the form as the validation container. This is necessary because in some cases, such as wizard,
+     the value of the 'data-dp-validate-if' attribute may refer to a different fieldset.
+     By using the form, this condition can be found throughout the entire form.
+     */
     const form = validationContainer.tagName === 'FIELDSET' && validationContainer.form ? validationContainer.form : validationContainer
 
     try {